# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: HackAtlantic ATS API
  version: 0.1.0
  description: Authoritative HTTP interface. Only implemented endpoints belong here.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/me:
    get:
      operationId: getCurrentUser
      summary: Resolve the authenticated Clerk user and local authorization roles
      tags: [Identity]
      security:
        - clerkBearer: []
      responses:
        "200":
          description: Current local user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "503":
          $ref: "#/components/responses/IdentityUnavailable"
  /v1/application-forms/current:
    get:
      operationId: getCurrentApplicationForm
      summary: Get the active open published application form
      tags: [Applications]
      security:
        - clerkBearer: []
      responses:
        "200":
          description: The current form that can receive a new application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentApplicationForm"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ApplicantRequired"
        "404":
          $ref: "#/components/responses/CurrentFormNotFound"
        "503":
          $ref: "#/components/responses/IdentityUnavailable"
  /v1/applications:
    post:
      operationId: createApplication
      summary: Start an empty draft application for the current form
      tags: [Applications]
      security:
        - clerkBearer: []
      responses:
        "200":
          description: Existing or newly created draft application for the current cycle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ApplicantRequired"
        "404":
          $ref: "#/components/responses/CurrentFormNotFound"
        "503":
          $ref: "#/components/responses/IdentityUnavailable"
  /v1/applications/mine:
    get:
      operationId: listMyApplications
      summary: List applications owned by the authenticated applicant
      tags: [Applications]
      security:
        - clerkBearer: []
      responses:
        "200":
          description: Applications owned by the applicant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationList"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ApplicantRequired"
        "503":
          $ref: "#/components/responses/IdentityUnavailable"
  /v1/applications/{applicationId}/draft:
    put:
      operationId: saveApplicationDraft
      summary: Replace every answer in an owned draft application
      tags: [Applications]
      security:
        - clerkBearer: []
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveApplicationDraftRequest"
      responses:
        "200":
          description: Updated application draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ApplicantRequired"
        "404":
          $ref: "#/components/responses/ApplicationNotFound"
        "409":
          $ref: "#/components/responses/ApplicationConflict"
        "422":
          $ref: "#/components/responses/InvalidAnswers"
        "503":
          $ref: "#/components/responses/IdentityUnavailable"
  /v1/applications/{applicationId}/submit:
    post:
      operationId: submitApplication
      summary: Submit a complete owned draft and queue its confirmation
      tags: [Applications]
      security:
        - clerkBearer: []
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitApplicationRequest"
      responses:
        "200":
          description: Submitted application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ApplicantRequired"
        "404":
          $ref: "#/components/responses/ApplicationNotFound"
        "409":
          $ref: "#/components/responses/ApplicationSubmissionConflict"
        "422":
          $ref: "#/components/responses/IncompleteApplication"
        "503":
          $ref: "#/components/responses/IdentityUnavailable"
  /v1/applications/{applicationId}/resume:
    get:
      operationId: getApplicationResume
      summary: Get metadata for the authenticated applicant's resume
      tags: [Applications]
      security:
        - clerkBearer: []
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
      responses:
        "200":
          description: Resume metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResume"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ApplicantRequired"
        "404":
          $ref: "#/components/responses/ApplicationNotFound"
    put:
      operationId: uploadApplicationResume
      summary: Upload or replace a PDF resume on an owned draft application
      tags: [Applications]
      security:
        - clerkBearer: []
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
        - name: X-File-Name
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/pdf:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Stored resume metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResume"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ApplicantRequired"
        "404":
          $ref: "#/components/responses/ApplicationNotFound"
        "413":
          description: Resume exceeds the 5 MiB limit
        "422":
          description: Body is not a valid PDF resume
  /v1/applications/{applicationId}/decision:
    get:
      operationId: getApplicantDecision
      summary: Get the caller's current released decision
      description: This applicant-safe response is absent until the current decision has been released. It never includes internal reasons or review content.
      tags: [Applications]
      security:
        - clerkBearer: []
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
      responses:
        "200":
          description: The applicant's released decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicantDecision"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ApplicantRequired"
        "404":
          $ref: "#/components/responses/DecisionNotFound"
        "503":
          $ref: "#/components/responses/IdentityUnavailable"
  /v1/applications/{applicationId}/rsvp:
    get:
      operationId: getApplicationRSVP
      summary: Get the caller's attendance response for their released acceptance
      tags: [Applications]
      security:
        - clerkBearer: []
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
      responses:
        "200":
          description: Current RSVP; pending means no response to this acceptance yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRSVP"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ApplicantRequired"
        "404":
          $ref: "#/components/responses/RSVPNotAvailable"
        "503":
          $ref: "#/components/responses/IdentityUnavailable"
    put:
      operationId: respondToRSVP
      summary: Confirm or decline attendance without changing admission or pass access
      description: Only the owner of a current released acceptance can respond. Send the decisionId and lockVersion from GET. Repeating the current response is idempotent; conflicting changes or a replaced acceptance require reloading after a 409. Pending is read-only. Actual changes are audited atomically.
      tags: [Applications]
      security:
        - clerkBearer: []
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RespondToRSVP"
      responses:
        "200":
          description: Persisted attendance response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRSVP"
        "400":
          description: Malformed payload or missing lockVersion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ApplicantRequired"
        "404":
          $ref: "#/components/responses/RSVPNotAvailable"
        "409":
          description: Conflicting response version or replaced acceptance; reload before retrying
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Invalid decisionId, status, or negative lockVersion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          $ref: "#/components/responses/IdentityUnavailable"
  /v1/admin/applications:
    get:
      operationId: listOrganizerApplications
      summary: Search and filter internal application projections
      tags: [Organizer]
      security:
        - clerkBearer: []
      parameters:
        - name: rsvp
          in: query
          required: false
          description: Filters current released acceptances only; pending means no RSVP yet
          schema:
            type: string
            enum: [pending, confirmed, declined]
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [submitted, accepted, waitlisted, rejected]
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Server-filtered organizer application queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizerApplicationList"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/OrganizerRequired"
        "422":
          $ref: "#/components/responses/InvalidFilter"
        "503":
          $ref: "#/components/responses/IdentityUnavailable"
  /v1/admin/applications/{applicationId}:
    get:
      operationId: getOrganizerApplication
      summary: Get an internal organizer application projection
      tags: [Organizer]
      security:
        - clerkBearer: []
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
      responses:
        "200":
          description: Internal application detail for organizers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizerApplication"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/OrganizerRequired"
        "404":
          $ref: "#/components/responses/ApplicationNotFound"
        "503":
          $ref: "#/components/responses/IdentityUnavailable"
  /v1/admin/applications/{applicationId}/resume:
    get:
      operationId: getAdminApplicationResume
      summary: Stream an application resume inline to an authorized admin
      tags: [Organizer]
      security:
        - clerkBearer: []
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
      responses:
        "200":
          description: Applicant resume PDF
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/OrganizerRequired"
        "404":
          $ref: "#/components/responses/ApplicationNotFound"
  /v1/admin/applications/{applicationId}/assignments:
    post:
      operationId: assignReviewer
      summary: Add auditable reviewer queue metadata to a submitted application
      tags: [Organizer]
      security:
        - clerkBearer: []
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignReviewerRequest"
      responses:
        "200":
          description: Reviewer assignment is recorded or already existed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewerAssignmentResult"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/OrganizerRequired"
        "404":
          $ref: "#/components/responses/ApplicationNotFound"
        "409":
          $ref: "#/components/responses/ApplicationNotSubmitted"
        "422":
          $ref: "#/components/responses/InvalidReviewer"
        "503":
          $ref: "#/components/responses/IdentityUnavailable"
  /v1/admin/applications/{applicationId}/decisions:
    post:
      operationId: recordDecision
      summary: Append an internal decision for a submitted application
      description: The application cache is updated atomically. A newer decision supersedes the previous current decision without mutating its history.
      tags: [Organizer]
      security:
        - clerkBearer: []
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordDecisionRequest"
      responses:
        "201":
          description: Appended internal decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizerDecision"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/OrganizerRequired"
        "404":
          $ref: "#/components/responses/ApplicationNotFound"
        "409":
          $ref: "#/components/responses/ApplicationNotSubmitted"
        "422":
          $ref: "#/components/responses/InvalidDecision"
        "503":
          $ref: "#/components/responses/IdentityUnavailable"
  /v1/admin/decisions/{decisionId}/release:
    post:
      operationId: releaseDecision
      summary: Release the current decision to its applicant
      description: Release is retry-safe and transactionally queues one decision email without internal decision or review data.
      tags: [Organizer]
      security:
        - clerkBearer: []
      parameters:
        - $ref: "#/components/parameters/DecisionID"
      responses:
        "200":
          description: Released decision or an idempotent release result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizerDecision"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/OrganizerRequired"
        "404":
          $ref: "#/components/responses/DecisionNotFound"
        "503":
          $ref: "#/components/responses/IdentityUnavailable"
  /v1/admin/users/{userId}/roles/reviewer:
    put:
      operationId: grantReviewerRole
      summary: Grant reviewer access and audit the role mutation
      tags: [Organizer]
      security:
        - clerkBearer: []
      parameters:
        - $ref: "#/components/parameters/UserID"
      responses:
        "204":
          description: Reviewer role is granted or already existed
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/OrganizerRequired"
        "404":
          $ref: "#/components/responses/UserNotFound"
        "503":
          $ref: "#/components/responses/IdentityUnavailable"
  /v1/admin/users/scanner-access/lookup:
    post:
      operationId: lookupScannerUser
      summary: Find an existing volunteer by exact verified primary email (admin only)
      description: Read-only lookup. Emails are sent in the body to avoid URL logging. Refuses ambiguous or stale identities; never creates an account or grants access.
      tags: [Organizer]
      security:
        - clerkBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email]
              properties:
                email: { type: string, format: email, maxLength: 254 }
      responses:
        "200":
          description: Matching account and scanner-access state; Cache-Control no-store
          content:
            application/json:
              schema:
                type: object
                required: [id, email, displayName, scannerAccess, canManage]
                properties:
                  id: { type: string, format: uuid }
                  email: { type: string, format: email }
                  displayName: { type: [string, "null"] }
                  scannerAccess: { type: boolean }
                  canManage: { type: boolean, description: False for admins and the caller's own account }
        "400":
          description: Malformed request
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/OrganizerRequired"
        "404":
          $ref: "#/components/responses/UserNotFound"
        "409":
          description: Multiple local accounts match the email; no account selected
        "422":
          description: Invalid email address
        "503":
          $ref: "#/components/responses/IdentityUnavailable"
  /v1/admin/users/{userId}/roles/scanner:
    put:
      operationId: grantScannerRole
      summary: Grant scanner access and audit the role mutation
      tags: [Organizer]
      security:
        - clerkBearer: []
      parameters:
        - $ref: "#/components/parameters/UserID"
      responses:
        "204":
          description: Scanner role is granted or already existed
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/OrganizerRequired"
        "404":
          $ref: "#/components/responses/UserNotFound"
        "503":
          $ref: "#/components/responses/IdentityUnavailable"
    delete:
      operationId: revokeScannerRole
      summary: Revoke scanner access and audit the role mutation
      tags: [Organizer]
      security:
        - clerkBearer: []
      parameters:
        - $ref: "#/components/parameters/UserID"
      responses:
        "204":
          description: Scanner role is revoked or already absent
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/OrganizerRequired"
        "404":
          $ref: "#/components/responses/UserNotFound"
        "503":
          $ref: "#/components/responses/IdentityUnavailable"
...
components:
  securitySchemes:
    clerkBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthenticated:
      description: Authentication was missing or invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ApplicantRequired:
      description: The authenticated identity does not have applicant access.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    CurrentFormNotFound:
      description: There is no active, open, published application form.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    IdentityUnavailable:
      description: Verified identity reconciliation is temporarily unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    CurrentUser:
      type: object
      additionalProperties: false
      required: [id, email, displayName, roles]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: [string, "null"]
        roles:
          type: array
          items:
            type: string
            enum: [applicant, admin, scanner]
...
    CurrentApplicationForm:
      type: object
      additionalProperties: false
      required: [id, cycleId, version, resumeRequired, questions]
      properties:
        id:
          type: string
          format: uuid
        cycleId:
          type: string
          format: uuid
        version:
          type: integer
          minimum: 1
        resumeRequired:
          type: boolean
        resumeAfterQuestionKey:
          type: string
          minLength: 1
        questions:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/FormQuestion"
...